+++ tests/ip-endpoint.nix
@@ -1,6 +1,6 @@
 { harness }:
 let
-  ep = import ../lib/endpoint.nix;
+  ep = import ../lib/ip-endpoint.nix;
   ipv4 = import ../lib/ipv4.nix;
   ipv6 = import ../lib/ipv6.nix;
   port = import ../lib/port.nix;